pyproject.toml
[project]
name = "cosmetic-sales-segementation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hdbscan>=0.8.41",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "numba>=0.63.1",
    "pandas>=2.3.3",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "umap-learn>=0.5.11",
]
